+++ packages/shared_preferences/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Increase the minimum Flutter version to 3.3.
 * Update shared_preferences to 2.2.0.
 * Update shared_preferences_interface to 2.3.0.
+* Add `clearWithParameters` and `getAllWithParameters`.
 * Update integration_test.
 
 ## 2.1.0